extract files and create index.ts
