+++ packages/electron-publish/package.json
@@ -1,7 +1,7 @@
 {
   "name": "electron-publish",
   "version": "26.0.0-alpha.1",
-  "main": "out/publisher.js",
+  "main": "out/index.js",
   "author": "Vladimir Krivosheev",
   "license": "MIT",
   "repository": {
