+++ packages/electron-publish/src/BitbucketPublisher.ts
@@ -1,7 +1,8 @@
 import { Arch, InvalidConfigurationError, isEmptyOrSpaces, log } from "builder-util"
 import { httpExecutor } from "builder-util"
 import { ClientRequest, RequestOptions } from "http"
-import { HttpPublisher, PublishContext } from "./publisher"
+import { PublishContext } from "./publisher"
+import { HttpPublisher } from "./HttpPublisher"
 import { BitbucketOptions } from "builder-util-runtime/out/publishOptions"
 import { configureRequestOptions, HttpExecutor } from "builder-util-runtime"
 import * as FormData from "form-data"
