+++ packages/electron-publish/src/KeygenPublisher.ts
@@ -1,7 +1,8 @@
 import { Arch, InvalidConfigurationError, log, isEmptyOrSpaces } from "builder-util"
 import { httpExecutor } from "builder-util"
 import { ClientRequest, RequestOptions } from "http"
-import { HttpPublisher, PublishContext } from "./publisher"
+import { PublishContext } from "./publisher"
+import { HttpPublisher } from "./HttpPublisher"
 import { KeygenOptions } from "builder-util-runtime/out/publishOptions"
 import { configureRequestOptions, HttpExecutor, parseJson } from "builder-util-runtime"
 import { getCompleteExtname } from "builder-util/out/filename"
