+++ packages/electron-publish/src/gitHubPublisher.ts
@@ -5,7 +5,9 @@ import { ClientRequest } from "http"
 import { Lazy } from "lazy-val"
 import * as mime from "mime"
 import { parse as parseUrl, UrlWithStringQuery } from "url"
-import { getCiTag, HttpPublisher, PublishContext, PublishOptions } from "./publisher"
+import { getCiTag, PublishContext } from "./publisher"
+import { HttpPublisher } from "./HttpPublisher"
+import { PublishOptions } from "./index"
 
 export interface Release {
   id: number
