+++ packages/electron-publish/src/index.ts
@@ -1,0 +1,10 @@
+export type PublishPolicy = "onTag" | "onTagOrDraft" | "always" | "never"
+
+export { ProgressCallback } from "./progress"
+
+export interface PublishOptions {
+  publish?: PublishPolicy | null
+}
+
+export { PublishContext, UploadTask, Publisher, getCiTag } from "./publisher"
+export { HttpPublisher } from "./HttpPublisher"
